realize_degree_sequence: update tests for extended API
