--- examples/simple/igraph_realize_degree_sequence.out
@@ -1,159 +1,0 @@
-3 2 2 1
-
-2 0
-1 0
-3 0
-2 1
-
-3 0
-2 0
-2 1
-1 0
-
-1 0
-2 0
-3 0
-2 1
-
-1 3 3 4 1 2 1 1 1 3
-
-9 3
-3 2
-3 1
-5 3
-9 2
-9 1
-2 1
-8 5
-7 6
-4 0
-
-8 3
-7 3
-6 1
-4 2
-9 0
-9 5
-5 2
-2 1
-9 3
-3 1
-
-3 0
-3 1
-2 1
-9 1
-3 2
-9 2
-5 3
-5 4
-9 6
-8 7
-
-2 0 3 2 2 2 2 3
-
-7 2
-7 6
-7 5
-4 2
-3 2
-4 0
-3 0
-6 5
-
-6 2
-7 6
-5 2
-7 5
-7 0
-3 2
-4 0
-4 3
-
-2 0
-7 0
-7 2
-3 2
-4 3
-5 4
-6 5
-7 6
-
-3 0 1 1 1 1 0 1
-2 1 0 2 2 1 0 0
-
-0 3
-0 4
-0 5
-3 0
-4 0
-5 4
-2 3
-7 1
-
-7 0
-2 3
-5 4
-3 0
-0 4
-0 3
-0 5
-4 1
-
-0 3
-0 4
-0 5
-2 0
-3 4
-4 3
-5 0
-7 1
-
-3 1 2 3 1 2 2
-2 2 1 2 3 2 2
-
-4 0
-3 4
-3 5
-3 6
-1 4
-0 1
-0 3
-0 5
-6 2
-6 1
-2 6
-2 3
-5 0
-5 4
-
-2 4
-2 0
-0 3
-0 5
-0 6
-6 4
-6 1
-1 3
-3 5
-3 4
-3 1
-4 6
-5 0
-5 2
-
-0 4
-0 3
-0 5
-1 6
-2 4
-2 1
-3 0
-3 6
-3 5
-4 0
-5 4
-5 3
-6 1
-6 2
-
